chore: lint fixed and refactor
